Clean up plugin.
Add proper scopes in test, autoImport and generally clean it up.
