+++ 0.13/hoarder/src/main/sbt_0.13/org.romanowski.hoarder/core/HoarderEngine.scala
@@ -14,7 +14,7 @@ import java.nio.file.Files
 import java.nio.file.Path
 import java.util
 
-import org.romanowski.HoarderSettings.CacheSetup
+import org.romanowski.HoarderPlugin.CacheSetup
 import org.romanowski.hoarder.core.SbtTypes.CompilationResult
 import org.romanowski.hoarder.core.SbtTypes.PreviousCompilationResult
 import sbt.PathFinder
