+++ 0.13/hoarder/src/main/scala/org/romanowski/hoarder/actions/StashKeys.scala
@@ -1,0 +1,19 @@
+package org.romanowski.hoarder.actions
+
+import sbt._
+
+trait StashKeys {
+  val stashKey = InputKey[Unit]("stash",
+    "Export results of current compilation to global location.")
+  val stashApplyKey = InputKey[Unit]("stashApply",
+    "Load exported compilation results from global location.")
+  val stashCleanKey = InputKey[Unit]("stashClean",
+    "Clean exported compilation results from global location.")
+
+  val defaultVersionLabel = SettingKey[String]("hoarder:defaultVersionLabel",
+    "Default version label for stash/stashApply")
+  val defaultProjectLabel = SettingKey[String]("hoarder:defaultProjectLabel",
+    "Default root project label for stash/stashApply")
+  val globalStashLocation = TaskKey[File]("hoarder:globalStashLocation",
+    "Root directory where exported artifacts are kept.")
+}
