+++ 0.13/hoarder/src/main/scala/org/romanowski/hoarder/core/HoarderEngineCommon.scala
@@ -8,7 +8,7 @@ package org.romanowski.hoarder.core
 
 import java.nio.file.Path
 
-import org.romanowski.HoarderSettings._
+import org.romanowski.HoarderPlugin._
 import org.romanowski.hoarder.core.SbtTypes.CompilationResult
 import org.romanowski.hoarder.core.SbtTypes.PreviousCompilationResult
