+++ 0.13/hoarderTests/src/sbt-test/cached-ci/multiProjectTest/build.sbt
@@ -1,10 +1,8 @@
 import sbt.dsl._
 
-def settings = org.romanowski.hoarder.tests.PluginTests.testRecompilation
-
-lazy val baseProject = project settings(settings:_*)
-lazy val finalProject = project settings(settings:_*) dependsOn(leafProject)
-lazy val leafProject = project settings(settings:_*) dependsOn(baseProject)
+lazy val baseProject = project
+lazy val finalProject = project dependsOn(leafProject)
+lazy val leafProject = project dependsOn(baseProject)
 
 
 val root = project aggregate(leafProject, baseProject, finalProject)
