--- 0.13/hoarderTests/src/sbt-test/cached-ci/simpleTest/new-src/Test2.scala
@@ -1,7 +1,0 @@
-/*
- * Hoarder - Cached compilation plugin for sbt.
- * Copyright 2016 - 2017, Krzysztof Romanowski
- * This software is released under the terms written in LICENSE.
- */
-
-class Test2
