+++ 0.13/hoarderTests/src/sbt-test/stash/crossVersionTest/build.sbt
@@ -1,3 +1,1 @@
-org.romanowski.hoarder.tests.PluginTests.testRecompilation
-
 crossScalaVersions := Seq("2.10.6","2.11.8", "2.12.1")
