--- 0.13/hoarderTests/src/sbt-test/stash/labelsTests/build.sbt
@@ -1,1 +1,0 @@
-org.romanowski.hoarder.tests.PluginTests.testRecompilation
