+++ 0.13/hoarderTests/src/sbt-test/stash/readFromTextTest/build.sbt
@@ -1,3 +1,1 @@
-org.romanowski.hoarder.tests.PluginTests.testRecompilation
-
-org.romanowski.HoarderSettings.zipAnalysisFile := false
+zipHoarderAnalysisCache := false
